Move selection related models into "Selection" module
I found this works better to describe the domain then using the plural
"Ratings", since this in fact encapsulates a domain of the app rather
then grouping different types of ratings.
